[refact] refactored sign up page (included github sign up)
